client: introduce the connection ctx manager (#8940)
ref tikv/pd#8690

Previously, we used a `sync.Map` as a medium to propagate connection ctx updates between
the dispatcher and TSO client, which introduced a lot of redundant parameter passing and
made the logic less intuitive. This PR implements the same functionality using a common
connection ctx manager to simplify and reuse related code.

Signed-off-by: JmPotato <ghzpotato@gmail.com>